feat: Implement visual updates for Navbar and Hero section
This commit introduces several visual enhancements based on user feedback:

1.  **Static White Navbar:**
    - The navbar is now permanently white with a subtle box-shadow to ensure it is always visible and stands out.
    - All scroll-based color-changing logic has been removed from `Navbar.js` and `App.js`.
    - Text and link colors have been adjusted to dark charcoal for proper contrast.

2.  **Social Media Icons:**
    - Facebook and Instagram icons from `react-icons` have been added to the right side of the navbar.
    - These icons are linked to the respective social media pages and are hidden on mobile to maintain a clean layout.

3.  **Hero Video Overlay Effect:**
    - A dynamic overlay has been added to the hero section video.
    - The overlay starts with a very light `#545454` tint and gradually darkens as the user scrolls down the page, creating a smooth darkening effect.
    - This is controlled by a `useEffect` hook in `HeroSection.js` that listens to the scroll position.
