+++ package.json
@@ -12,6 +12,7 @@
     "express": "^4.19.2",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-icons": "^5.5.0",
     "react-router-dom": "^7.9.3",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
@@ -20,13 +21,14 @@
     "concurrently": "^8.2.2"
   },
   "scripts": {
-    "start": "node backend/server.js",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "server": "node backend/server.js",
-    "dev": "concurrently \"react-scripts start\" \"npm run server\""
+    "dev": "concurrently \"npm run start\" \"npm run server\""
   },
+  "proxy": "http://localhost:3001",
   "eslintConfig": {
     "extends": [
       "react-app",
