+++ src/components/HeroSection.css
@@ -33,7 +33,9 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(255, 255, 255, 0.1); /* Very subtle white overlay */
+  /* The background color is now controlled by an inline style in HeroSection.js */
+  /* This transition will smooth the opacity change on scroll */
+  transition: background-color 0.1s linear;
 }
 
 .hero-content {
