+++ src/components/Navbar.css
@@ -1,6 +1,6 @@
 /* src/components/Navbar.css */
 
-/* --- Default (Transparent) State --- */
+/* --- Static White Navbar --- */
 .main-navbar {
   position: fixed;
   top: 0;
@@ -11,15 +11,16 @@
   align-items: center;
   padding: 20px 40px;
   z-index: 1000;
-  background-color: transparent;
-  transition: background-color 0.4s ease, color 0.4s ease, box-shadow 0.4s ease;
+  background-color: var(--background-primary); /* Always white */
+  box-shadow: 0 2px 4px rgba(0,0,0,0.05); /* Subtle shadow */
 }
 
+/* --- Links and Logo --- */
 .main-navbar .navbar-logo a,
 .main-navbar .navbar-links > li > a {
-  color: var(--background-primary);
-  text-shadow: 0 0 8px rgba(0, 0, 0, 0.8);
-  transition: color 0.3s ease, text-shadow 0.3s ease;
+  color: var(--text-charcoal); /* Dark text for contrast */
+  text-shadow: none;
+  transition: color 0.3s ease;
 }
 
 .main-navbar .navbar-links > li > a {
@@ -31,40 +32,9 @@
   cursor: pointer;
 }
 
-/* --- Scrolled State --- */
-.main-navbar.navbar-scrolled {
-  background-color: #8f8d8d;
-  box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-}
-
-.main-navbar.navbar-scrolled .navbar-links a,
-.main-navbar.navbar-scrolled .navbar-logo a {
-  color: var(--background-primary);
-  text-shadow: none;
-}
-
-.main-navbar.navbar-scrolled .dropdown-menu {
-  background-color: #8f8d8d;
-}
-
-.main-navbar.navbar-scrolled .dropdown-menu a {
-  color: var(--background-primary);
-}
-
-.main-navbar.navbar-scrolled .dropdown-menu a:hover {
-  background-color: #333;
-}
-
-/* --- Admin Page State (Static White) --- */
-.main-navbar.navbar-admin {
-  background-color: var(--background-primary) !important;
-  box-shadow: 0 2px 4px rgba(0,0,0,0.05);
-}
-
-.main-navbar.navbar-admin .navbar-links a,
-.main-navbar.navbar-admin .navbar-logo a {
-  color: var(--text-charcoal) !important;
-  text-shadow: none !important;
+.main-navbar .navbar-links > li > a:hover,
+.main-navbar .navbar-logo a:hover {
+  color: var(--text-dark-gray);
 }
 
 /* --- General Styles --- */
@@ -105,8 +75,7 @@
   position: absolute;
   top: 100%;
   left: 0;
-  background-color: rgba(255, 255, 255, 0.9);
-  backdrop-filter: blur(5px);
+  background-color: var(--background-primary);
   box-shadow: 0 8px 16px rgba(0,0,0,0.1);
   list-style: none;
   padding: 10px 0;
@@ -132,6 +101,26 @@
   background-color: #f1f1f1;
 }
 
+/* --- Social Icons --- */
+.social-icons {
+  display: flex;
+  align-items: center;
+  gap: 18px;
+  margin-right: 15px;
+}
+
+.social-icons a {
+  color: var(--text-charcoal);
+  font-size: 22px; /* Icon size */
+  transition: color 0.3s ease, transform 0.3s ease;
+}
+
+.social-icons a:hover {
+  color: var(--text-dark-gray);
+  transform: scale(1.1);
+}
+
+
 /* --- Mobile Menu Styles --- */
 .mobile-menu-toggle {
   display: none;
@@ -148,12 +137,8 @@
   height: 3px;
   margin: 5px 0;
   transition: all 0.3s ease-in-out;
-  background-color: var(--background-primary); /* Default to white for hero */
+  background-color: var(--text-charcoal); /* Always dark */
 }
-
-/* Color for scrolled and admin states */
-.main-navbar.navbar-scrolled .mobile-menu-toggle span { background-color: var(--background-primary); }
-.main-navbar.navbar-admin .mobile-menu-toggle span { background-color: var(--text-charcoal) !important; }
 
 /* Hamburger to "X" animation */
 .mobile-menu-toggle.open span:nth-child(1) { transform: translateY(8px) rotate(45deg); }
@@ -171,6 +156,10 @@
     transform: none;
     text-align: left;
   }
+  .social-icons {
+    display: none; /* Hide social icons on mobile to save space */
+  }
+
   .mobile-menu-toggle { display: block; }
 
   .navbar-links {
@@ -201,11 +190,11 @@
   }
 
   .dropdown-menu {
-    display: none; /* Correctly hide by default */
+    display: none;
     position: static;
     box-shadow: none;
     background: transparent;
-    padding: 10px 0 10px 20px; /* Indent sub-items */
+    padding: 10px 0 10px 20px;
     min-width: unset;
     border-left: 2px solid #f0f0f0;
     margin-left: 10px;
@@ -216,8 +205,8 @@
   }
 
   .dropdown-menu.open {
-    display: block; /* Show on click */
-    max-height: 500px; /* Allow it to expand */
+    display: block;
+    max-height: 500px;
     transition: max-height 0.4s ease-in;
   }
 
@@ -228,7 +217,6 @@
   }
 }
 
-/* Prevent body scroll when mobile menu is open */
 body.no-scroll {
   overflow: hidden;
 }
