+++ src/index.css
@@ -3,8 +3,8 @@
 :root {
   /* Color Palette */
   --background-primary: #FFFFFF;
-  --text-charcoal: #0D0D0D;
-  --text-dark-gray: #555555;
+  --text-charcoal: #545454;
+  --text-dark-gray: #8f8d8d;
   --text-medium-gray: #444444;
   --border-light-gray: #CCCCCC;
   --input-background: #F5F5F5;
